pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "github-scraper"
version = "0.0.1"
description = "Generate professional CVs from GitHub profiles with comprehensive data scraping and analytics"
authors = [{name="Nikos Mavridis"}]
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "typer>=0.7",
    "PyGithub>=2.3",
    "Jinja2>=3.1",
    "httpx>=0.24",
]

[project.scripts]
github-scraper = "github_scraper.cli:app"

[project.optional-dependencies]
# YAML configuration support
yaml = [
    "PyYAML>=6.0",
]

# All optional features
all = [
    "PyYAML>=6.0",
]

# Development dependencies
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.6",
    "types-PyYAML>=6.0",
]